fix(sim): port loot-FFA into v0.16.0 interaction.ts + regen parity goldens
