+++ tests/parity/golden/hunter_pet.json
@@ -101,7 +101,7 @@
       "time": 1,
       "nextId": 392,
       "state": "c91a435c",
-      "events": "3008f6c4",
+      "events": "bf8da7fe",
       "rng": {
         "draws": 6,
         "digest": "fb26b191"
@@ -123,7 +123,7 @@
       "time": 3,
       "nextId": 392,
       "state": "97481452",
-      "events": "f325a859",
+      "events": "b98e45cd",
       "rng": {
         "draws": 96,
         "digest": "ee95214e"
@@ -145,7 +145,7 @@
       "time": 5,
       "nextId": 392,
       "state": "671fa3b4",
-      "events": "f25735b6",
+      "events": "6b3b0c7c",
       "rng": {
         "draws": 294,
         "digest": "98724024"
