+++ tests/parity/golden/inventory_vendor.json
@@ -35,7 +35,7 @@
       "tick": 0,
       "time": 0,
       "nextId": 389,
-      "state": "f3a390f2",
+      "state": "df8b348a",
       "events": "741638a5",
       "rng": {
         "draws": 0,
@@ -78,6 +78,7 @@
           "id": 388,
           "kind": "player",
           "level": 1,
+          "lootFfaTimer": "Infinity",
           "maxHp": 90,
           "maxResource": 100,
           "moveSpeed": 7,
@@ -117,7 +118,7 @@
       "tick": 0,
       "time": 0,
       "nextId": 389,
-      "state": "bfbaece0",
+      "state": "795b8668",
       "events": "21673dd2",
       "rng": {
         "draws": 0,
@@ -174,6 +175,7 @@
           "id": 388,
           "kind": "player",
           "level": 1,
+          "lootFfaTimer": "Infinity",
           "maxHp": 90,
           "maxResource": 100,
           "moveSpeed": 7,
@@ -213,7 +215,7 @@
       "tick": 0,
       "time": 0,
       "nextId": 389,
-      "state": "628a8256",
+      "state": "935b3ece",
       "events": "9df425a0",
       "rng": {
         "draws": 0,
@@ -275,6 +277,7 @@
           "id": 388,
           "kind": "player",
           "level": 1,
+          "lootFfaTimer": "Infinity",
           "maxHp": 120,
           "maxResource": 100,
           "moveSpeed": 7,
@@ -314,7 +317,7 @@
       "tick": 0,
       "time": 0,
       "nextId": 389,
-      "state": "96db94ae",
+      "state": "89208e1a",
       "events": "dbe7ae9a",
       "rng": {
         "draws": 0,
@@ -376,6 +379,7 @@
           "id": 388,
           "kind": "player",
           "level": 1,
+          "lootFfaTimer": "Infinity",
           "maxHp": 110,
           "maxResource": 100,
           "moveSpeed": 7,
@@ -415,7 +419,7 @@
       "tick": 0,
       "time": 0,
       "nextId": 389,
-      "state": "9d7aaca2",
+      "state": "fd5c251a",
       "events": "a0e9b547",
       "rng": {
         "draws": 0,
@@ -480,6 +484,7 @@
           "id": 388,
           "kind": "player",
           "level": 1,
+          "lootFfaTimer": "Infinity",
           "maxHp": 90,
           "maxResource": 100,
           "moveSpeed": 7,
@@ -519,7 +524,7 @@
       "tick": 0,
       "time": 0,
       "nextId": 389,
-      "state": "d0bfe856",
+      "state": "c9808616",
       "events": "db521cfd",
       "rng": {
         "draws": 0,
@@ -588,6 +593,7 @@
           "id": 388,
           "kind": "player",
           "level": 1,
+          "lootFfaTimer": "Infinity",
           "maxHp": 90,
           "maxResource": 100,
           "moveSpeed": 7,
@@ -628,7 +634,7 @@
       "tick": 0,
       "time": 0,
       "nextId": 389,
-      "state": "d0914b9a",
+      "state": "ac3192da",
       "events": "579898e6",
       "rng": {
         "draws": 0,
@@ -693,6 +699,7 @@
           "id": 388,
           "kind": "player",
           "level": 1,
+          "lootFfaTimer": "Infinity",
           "maxHp": 90,
           "maxResource": 100,
           "moveSpeed": 7,
@@ -733,7 +740,7 @@
       "tick": 0,
       "time": 0,
       "nextId": 389,
-      "state": "0104bf48",
+      "state": "48246930",
       "events": "f087f305",
       "rng": {
         "draws": 0,
@@ -810,6 +817,7 @@
           "id": 388,
           "kind": "player",
           "level": 1,
+          "lootFfaTimer": "Infinity",
           "maxHp": 210,
           "maxResource": 100,
           "moveSpeed": 7,
@@ -850,7 +858,7 @@
       "tick": 0,
       "time": 0,
       "nextId": 389,
-      "state": "0c1d0b56",
+      "state": "46e1009e",
       "events": "ef990a2a",
       "rng": {
         "draws": 0,
@@ -931,6 +939,7 @@
           "id": 388,
           "kind": "player",
           "level": 1,
+          "lootFfaTimer": "Infinity",
           "maxHp": 210,
           "maxResource": 100,
           "moveSpeed": 7,
@@ -971,7 +980,7 @@
       "tick": 0,
       "time": 0,
       "nextId": 389,
-      "state": "d957a7b9",
+      "state": "1d8ffcd1",
       "events": "b71a55c4",
       "rng": {
         "draws": 0,
@@ -1058,6 +1067,7 @@
           "id": 388,
           "kind": "player",
           "level": 1,
+          "lootFfaTimer": "Infinity",
           "maxHp": 210,
           "maxResource": 100,
           "moveSpeed": 7,
@@ -1098,7 +1108,7 @@
       "tick": 0,
       "time": 0,
       "nextId": 389,
-      "state": "4c62201a",
+      "state": "4e49b192",
       "events": "7521a204",
       "rng": {
         "draws": 0,
@@ -1185,6 +1195,7 @@
           "id": 388,
           "kind": "player",
           "level": 1,
+          "lootFfaTimer": "Infinity",
           "maxHp": 210,
           "maxResource": 100,
           "moveSpeed": 7,
@@ -1225,7 +1236,7 @@
       "tick": 0,
       "time": 0,
       "nextId": 389,
-      "state": "61d4f7ec",
+      "state": "f7b661e4",
       "events": "e277cc91",
       "rng": {
         "draws": 0,
@@ -1316,6 +1327,7 @@
           "id": 388,
           "kind": "player",
           "level": 1,
+          "lootFfaTimer": "Infinity",
           "maxHp": 210,
           "maxResource": 100,
           "moveSpeed": 7,
@@ -1356,7 +1368,7 @@
       "tick": 0,
       "time": 0,
       "nextId": 389,
-      "state": "61d4f7ec",
+      "state": "f7b661e4",
       "events": "741638a5",
       "rng": {
         "draws": 0,
@@ -1447,6 +1459,7 @@
           "id": 388,
           "kind": "player",
           "level": 1,
+          "lootFfaTimer": "Infinity",
           "maxHp": 210,
           "maxResource": 100,
           "moveSpeed": 7,
