+++ tests/parity/golden/quest_link_abandon.json
@@ -29,7 +29,7 @@
       "tick": 0,
       "time": 0,
       "nextId": 390,
-      "state": "0be0c8a0",
+      "state": "763c0ad4",
       "events": "b36157a1",
       "rng": {
         "draws": 0,
@@ -90,6 +90,7 @@
           "id": 388,
           "kind": "player",
           "level": 1,
+          "lootFfaTimer": "Infinity",
           "maxHp": 90,
           "maxResource": 100,
           "moveSpeed": 7,
@@ -136,6 +137,7 @@
           "id": 389,
           "kind": "player",
           "level": 1,
+          "lootFfaTimer": "Infinity",
           "maxHp": 54,
           "maxResource": 195,
           "moveSpeed": 7,
@@ -176,7 +178,7 @@
       "tick": 0,
       "time": 0,
       "nextId": 390,
-      "state": "0be0c8a0",
+      "state": "763c0ad4",
       "events": "741638a5",
       "rng": {
         "draws": 0,
@@ -237,6 +239,7 @@
           "id": 388,
           "kind": "player",
           "level": 1,
+          "lootFfaTimer": "Infinity",
           "maxHp": 90,
           "maxResource": 100,
           "moveSpeed": 7,
@@ -283,6 +286,7 @@
           "id": 389,
           "kind": "player",
           "level": 1,
+          "lootFfaTimer": "Infinity",
           "maxHp": 54,
           "maxResource": 195,
           "moveSpeed": 7,
@@ -323,7 +327,7 @@
       "tick": 0,
       "time": 0,
       "nextId": 390,
-      "state": "0be0c8a0",
+      "state": "763c0ad4",
       "events": "2873f427",
       "rng": {
         "draws": 0,
@@ -384,6 +388,7 @@
           "id": 388,
           "kind": "player",
           "level": 1,
+          "lootFfaTimer": "Infinity",
           "maxHp": 90,
           "maxResource": 100,
           "moveSpeed": 7,
@@ -430,6 +435,7 @@
           "id": 389,
           "kind": "player",
           "level": 1,
+          "lootFfaTimer": "Infinity",
           "maxHp": 54,
           "maxResource": 195,
           "moveSpeed": 7,
@@ -470,7 +476,7 @@
       "tick": 0,
       "time": 0,
       "nextId": 390,
-      "state": "b1f86152",
+      "state": "ada27dde",
       "events": "77206bf1",
       "rng": {
         "draws": 0,
@@ -543,6 +549,7 @@
           "id": 388,
           "kind": "player",
           "level": 1,
+          "lootFfaTimer": "Infinity",
           "maxHp": 90,
           "maxResource": 100,
           "moveSpeed": 7,
@@ -589,6 +596,7 @@
           "id": 389,
           "kind": "player",
           "level": 1,
+          "lootFfaTimer": "Infinity",
           "maxHp": 54,
           "maxResource": 195,
           "moveSpeed": 7,
@@ -629,7 +637,7 @@
       "tick": 0,
       "time": 0,
       "nextId": 390,
-      "state": "0be0c8a0",
+      "state": "763c0ad4",
       "events": "72a2d48f",
       "rng": {
         "draws": 0,
@@ -690,6 +698,7 @@
           "id": 388,
           "kind": "player",
           "level": 1,
+          "lootFfaTimer": "Infinity",
           "maxHp": 90,
           "maxResource": 100,
           "moveSpeed": 7,
@@ -736,6 +745,7 @@
           "id": 389,
           "kind": "player",
           "level": 1,
+          "lootFfaTimer": "Infinity",
           "maxHp": 54,
           "maxResource": 195,
           "moveSpeed": 7,
@@ -776,7 +786,7 @@
       "tick": 2,
       "time": 0.1,
       "nextId": 390,
-      "state": "8ae44b64",
+      "state": "71089b3c",
       "events": "741638a5",
       "rng": {
         "draws": 0,
@@ -837,6 +847,7 @@
           "id": 388,
           "kind": "player",
           "level": 1,
+          "lootFfaTimer": "Infinity",
           "maxHp": 90,
           "maxResource": 100,
           "moveSpeed": 7,
@@ -883,6 +894,7 @@
           "id": 389,
           "kind": "player",
           "level": 1,
+          "lootFfaTimer": "Infinity",
           "maxHp": 54,
           "maxResource": 195,
           "moveSpeed": 7,
